Rename - Package and Main App
패키지, 스프링 부트 메인 애플리케이션의 이름을 변경
