+++ src/main/java/me/kevinntech/onlineshop/Application.java
@@ -1,13 +1,13 @@
-package me.kevinntech.onlineshopreboot;
+package me.kevinntech.onlineshop;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class OnlineShopRebootApplication {
+public class Application {
 
 	public static void main(String[] args) {
-		SpringApplication.run(OnlineShopRebootApplication.class, args);
+		SpringApplication.run(Application.class, args);
 	}
 
 }
